refactor(errors): move UnusedRuleWarning to lib/errors
Error and warning classes belong in lib/errors regardless of which plugin
raises them. Simplify the guide's hints section to the reusable parts: where
the class lives, that severity comes from the existing hints reporting, and
how to tell whether a diagnostic needs a serializer.
