+++ lib/errors/UnusedRuleWarning.js
@@ -5,7 +5,7 @@
 
 "use strict";
 
-const WebpackError = require("../errors/WebpackError");
+const WebpackError = require("./WebpackError");
 
 class UnusedRuleWarning extends WebpackError {
 	/**
