+++ lib/performance/UnusedRuleWarningPlugin.js
@@ -5,7 +5,7 @@
 
 "use strict";
 
-const UnusedRuleWarning = require("./UnusedRuleWarning");
+const UnusedRuleWarning = require("../errors/UnusedRuleWarning");
 
 /** @typedef {import("../../declarations/WebpackOptions").RuleSetRule} RuleSetRule */
 /** @typedef {import("../../declarations/WebpackOptions").PerformanceOptions} PerformanceOptions */
